chasquid: Check result of initial os.Chdir
The daemon attempts to change to the config directory on startup, for
security and convenience.

We currently don't check if this works, which is not a big deal since it
will just fail later on when it can't find the files. However, it makes
things more awkward to debug, so this patch adds an explicit check.