travis: Enable Go modules in 1.11
Currently the modules are ignored in the Go 1.11 build, because the
files are within $GOPATH.

This causes problems when some dependencies are updated in
backwards-incompatible ways, and assuming that Go modules are being
used. In particular, the new protobuf release caused this problem which
was caught by the automated builds:
https://travis-ci.org/github/albertito/chasquid/jobs/674701956.

This patch enables Go modules in 1.11 builds.

Thanks to Jonas Seydel (thor77) for the help investigating and finding a
fix for this problem.